* New dependencies:     * `@novivia/babel`
* Updated dependencies:
    * `@novivia/build-module` to v0.5.
    * `babel-jest` to v17.
    * `jest` to v17.

* Removed dependencies:
    * `babel-plugin-add-module-exports`
    * `babel-plugin-transform-decorators-legacy`
    * `babel-polyfill`
    * `babel-preset-latest`
    * `babel-preset-react`
    * `babel-preset-stage-0`
    * `babel-register`
