+++ CHANGELOG.md
@@ -1,5 +1,25 @@
 # Versions
 
+## v0.1.0 - (02/12/2016)
+
+* New dependencies:
+    * `@novivia/babel`
+
+* Updated dependencies:
+    * `@novivia/build-module` to v0.5.
+    * `babel-jest` to v17.
+    * `jest` to v17.
+
+* Removed dependencies:
+    * `babel-plugin-add-module-exports`
+    * `babel-plugin-transform-decorators-legacy`
+    * `babel-polyfill`
+    * `babel-preset-latest`
+    * `babel-preset-react`
+    * `babel-preset-stage-0`
+    * `babel-register`
+
+
 ## v0.0.4 - (02/11/2016)
 
 * Now ignoring all `__tests__` subdirectories, not just `__fixtures__`.
