+++ lib/transform.js
@@ -5,5 +5,5 @@
 
 /* eslint-disable comma-dangle */
 module.exports = require("babel-jest").createTransformer(
-  require("../babelConfig")
+  require("@novivia/babel/config/backend")()
 );
