+++ package.json5
@@ -18,15 +18,22 @@
     "url": "git+ssh://git.novivia.com/nov/tools.tester"
   },
   "scripts": {
-    "build": "build-module build -b babelConfig.js -b cli.js -b monkeyPatch.js -p cliWrapper",
+    "build": "novivia-builder build",
     "lint": "novivia-linter lint",
     "open-source": "open-source",
-    "pub": "build-module publish",
-    "release": "build-module release",
+    "pub": "novivia-builder publish",
+    "release": "novivia-builder release",
     "test": "./cliWrapper",
   },
   "bin": {
     "novivia-tester": "./cliWrapper",
+  },
+  "novivia-builder": {
+    "babelPatterns": [
+      "cli.js",
+      "monkeyPatch.js",
+    ],
+    "packagePatterns": ["cliWrapper"],
   },
   "novivia-linter": {
     "targets": {
@@ -39,28 +46,26 @@
   "novivia-tester": {
     "coverageLocations": [
       "!cli.js",
+      "!lib/cli.js",
       "!monkeyPatch.js",
     ],
   },
+  "peerDependencies": {
+    "babel-polyfill": ">= 6.16.0",
+  },
   "devDependencies": {
     // Novivia dependencies.
-    "@novivia/build-module": "^0.4.0",
-    "@novivia/linter": "^1.6.0",
+    "@novivia/build-module": "^0.5.1",
+    "@novivia/linter": "^1.9.0",
     "@novivia/open-sourcer": "^0.1.1",
-
-    // npm dependencies.
-    "babel-register": "^6.18.0",
   },
   "dependencies": {
+    // Novivia dependencies.
+    "@novivia/babel": "0.0.5",
+
     // npm dependencies.
-    "babel-jest": "^16.0.0",
-    "babel-plugin-add-module-exports": "^0.2.1",
-    "babel-plugin-transform-decorators-legacy": "^1.3.4",
-    "babel-polyfill": "^6.16.0",
-    "babel-preset-latest": "^6.16.0",
-    "babel-preset-react": "^6.16.0",
-    "babel-preset-stage-0": "^6.16.0",
-    "jest": "^16.0.2",
+    "babel-jest": "^17.0.2",
+    "jest": "^17.0.3",
     "path-exists": "^3.0.0",
     "pkginfo-json5": "^0.5.0",
   },
